Code dump
This consists of all the code I wrote tonight.  It's good enough to
calculate the Elo ratings of players, but it doesn't yet have any
proper output functionality.  I currently just test it by running

./bin/generate_html | sort -hr

which outputs a sorted list of players by Elo rating.  I'll probably
get around to implementing the actual HTML output soon, but for now
this is enough to be interesting.